requirements.txt
gradio==4.12.0
together
python-dotenv
openai
SpeechRecognition
PyAudio
soundfile